삭제를 post처리

+++ metabbs/app/admin/controllers/board/delete.php
@@ -1,5 +1,7 @@
 <?php
-$board->delete();
-Flash::set('Board has been deleted.');
-redirect_to(url_admin_for('board'));
+if (is_post()) {
+	$board->delete();
+	Flash::set('Board has been deleted.');
+	redirect_to(url_admin_for('board'));
+}
 ?>
