+++ metabbs/app/admin/controllers/user/delete.php
@@ -1,5 +1,7 @@
 <?
-$user = User::find($params['id']);
-$user->delete();
-redirect_to(url_admin_for('user', null, array('page' => $_GET['page'])));
+if (is_post()) {
+	$user = User::find($params['id']);
+	$user->delete();
+	redirect_to(url_admin_for('user', null, array('page' => $_GET['page'])));
+}
 ?>
